Completed merge and added login page
