+++ frontend/src/App.vue
@@ -1,237 +1,129 @@
-export const backurl = process.env.VUE_APP_BACKEND_URL
 <template>
-  <div class="Chat">
-    <div class="channels-list">
-      <div v-if="side_info === 0" v-for="channel in channels" :key="channel.id" :class="['channel', { 'selected': channel.id === selected_channel }]" @click="chooseChannel(channel.id)">
-        {{ channel.channel_name }}
-      </div>
-      <div v-if="side_info === 1" v-for="user in users" :key="user.id" class="user">
-        {{ user.nick }}
-      </div>
-      <div v-if="side_info === 2" class="searchcontainer">
-        <form @submit.prevent="search">
-          <div class="search-input-container">
-            <button type="back" class="back-button" @click="side_info = 0"></button>
-            <input type="text" placeholder="Search..." class="input-field">
-            <button type="submit" class="send-button">Go</button>
-          </div>
-         </form>
-      </div>
-      <div class="button-container" v-if="side_info !== 2" >
-        <button class="channel-button bar-button" @click="getChannels()"></button>
-        <button class="people-button bar-button" @click="getUsers()"></button>
-        <button class="new-button bar-button" @click="createChannel()"></button>
-        <button class="search-button bar-button" @click="search()"></button>
-      </div>
+    <header v-if="islogged">
+      <nav>
+        <RouterLink to="/">Pong</RouterLink>
+        <RouterLink to="/chat">Chat</RouterLink>
+        <RouterLink to="/leaderboard">Leaderboard</RouterLink>
+        <RouterLink to="/profile">User profile</RouterLink>
+      </nav>
+    </header>
+    <div v-if="islogged">
+      <RouterView />
     </div>
-    <div id="chat-container" ref="chatContainer">
-      <div id="msg-container" ref="msgsContainer">
-        <div v-for="message in messages" :key="message.id" :class="[getMessageClass(message.author.nick), 'message']">
-          <strong>[{{ message.author?.nick }}]:</strong> {{ message.message }}
-          <div class="message-time">{{ formatTime(message.time) }}</div>
-        </div>
+    <div class="loginpage" v-else>
+      <div class="combo">
+        <Login />
+        <!-- Add this to Login and emit  -->
+        <button class="button" @click="test" >Login</button>
       </div>
-      <div class="msg-input">
-        <form @submit.prevent="sendMessage">
-          <input v-model="messageText" placeholder="Message" class="input-field">
-          <button type="submit" class="send-button">Send</button>
-        </form>
-      </div>
-    </div>
-  </div>
+    </div> 
 </template>
 
-<script setup>
-import { io } from 'socket.io-client'
-import { ref, onBeforeMount, watch, nextTick } from 'vue';
+<script setup lang="ts">
+import { RouterLink, RouterView } from 'vue-router'
+import Login from "./components/Login.vue"
+import { ref } from 'vue';
 
-const socket = io(process.env.VUE_APP_BACKEND_URL);
-const msgsContainer = ref(null);
-const messageText = ref('');
-const messages = ref([]);
-const channels = ref([]);
-const users = ref([]);
-let selected_channel = null;
-let side_info = ref(0);
+let islogged = ref(false);
 
-const check_user = async () => {
-try {
-  let url = process.env.VUE_APP_BACKEND_URL + '/users/getUsers/' + localStorage.name
-  const response = await fetch(url);
-  if (response.ok) {
-    const data = await response.json();  
-    localStorage.id=data.id
-  } else {
-    console.log('Error:', response.status);
-    window.alert("User Doesn't exist")
-  }
-} catch (error) {
-  console.log('Error:', error);
+function test()
+{
+  islogged.value=true
 }
-};
-
-//https://developer.mozilla.org/en-US/docs/Web/API/URLSearchParams/set
-let params = new URLSearchParams(document.location.search);
-let name = params.get("username"); // is the string "Jonathan"
-let intra_nick= params.get("intra_nick"); // is the number 18
-if(name && intra_nick){
-  localStorage.name=name;
-  localStorage.intra_nick=intra_nick;
-  check_user()
-}
-else{
-  window.alert("Please set username + intra:nick")
-  window.history.replaceState(null, '', '?username=PLEASE_SET&intra_nick=PLEASE_SET');
-}
-
-
-const getMessageClass = (author) => {
-  if (author == localStorage.name) {
-    return 'message-sent';
-  }
-  return 'message-received';
-};
-
-
-const getMessages = async () => {
-  try {
-    let url = process.env.VUE_APP_BACKEND_URL + '/chat/msginchannel/' + selected_channel
-    const response = await fetch(url);
-    if (response.ok) {
-      const data = await response.json();
-      messages.value = data;
-      scrollToBottom();
-    } else {
-      console.log('Error:', response.status);
-    }
-  } catch (error) {
-    console.log('Error:', error);
-  }
-};
-
-const getUsers = async () => {
-side_info.value = 1;
-try {
-  let url = process.env.VUE_APP_BACKEND_URL + '/users/getUsers'
-  const response = await fetch(url);
-  if (response.ok) {
-    const data = await response.json();
-    users.value = data;
-  } else {
-    console.log('Error:', response.status);
-  }
-} catch (error) {
-  console.log('Error:', error);
-}
-};
-
-
-const getChannels = async () => {
-side_info.value = 0;
-try {
-  let url = process.env.VUE_APP_BACKEND_URL + '/channels/all'
-  const response = await fetch(url);
-  if (response.ok) {
-    const data = await response.json();
-    channels.value = data;
-  } else {
-    console.log('Error:', response.status);
-  }
-} catch (error) {
-  console.log('Error:', error);
-}
-};
-
-const scrollToBottom = () => {
-  try {
-    nextTick(() => {
-      let container = msgsContainer.value;
-      container.scrollTop = container.scrollHeight;
-    });
-  } catch (error) {
-    console.log('Error:', error);
-  }
-};
-
-const formatTime = (timestamp) => {
-  const currentTime = new Date();
-  const messageTime = new Date(timestamp);
-  const timeDiffMinutes = Math.floor((currentTime - messageTime) / (1000 * 60));
-
-  if (timeDiffMinutes < 1) {
-    return 'Just now';
-  } else if (timeDiffMinutes < 60) {
-    return `${timeDiffMinutes} mins ago`;
-  } else if (timeDiffMinutes < 1440) {
-    const hours = Math.floor(timeDiffMinutes / 60);
-    return `${hours} ${hours === 1 ? 'hour' : 'hours'} ago`;
-  } else {
-    const days = Math.floor(timeDiffMinutes / 1440);
-    return `${days} ${days === 1 ? 'day' : 'days'} ago`;
-  }
-};
-
-
-
-const sendMessage = () => {
-  if (messageText.value == '' || selected_channel == null)
-    return window.alert("Error: Message cannot be empty. Please enter a valid message or join channel before sending ");
-  socket.emit('sendMessage', { authorId:localStorage.id, message: messageText.value, channelId: selected_channel })
-  console.log(messageText)
-  messageText.value = '';
-}
-
-const search = () => {
-  side_info.value = 2;
-}
-
-const chooseChannel = (channel)=>{
-  selected_channel = channel;
-  getMessages();
-}
-
-const createChannel = async () => {
-  let channel_name = window.prompt("Channel Name");
-  try {
-    let url = process.env.VUE_APP_BACKEND_URL + '/channels/create'
-    const response = await fetch(url, {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify({ type: '1', channel_name: channel_name }),
-    });
-    if (response.ok) {
-      const data = await response.json();
-      await getChannels();
-      chooseChannel(data.id);
-      console.log(data.id);
-    } else {
-      console.log('Error:', response.status);
-    }
-  } catch (error) {
-    console.log('Error:', error);
-  }
-};
-
-onBeforeMount(() => {
-  getChannels();
-});
-
-socket.on('recMessage', message => {
-  getMessages()
-});
-
-watch(messages, () => {
-  scrollToBottom();
-});
-
 
 </script>
 
+<style scoped>
+header {
+  line-height: 1.5;
+  max-height: 100vh;
+  display: flex;
+  margin-right: 2vw;
+}
 
-<style>
-  @import './App.css';
+nav {
+  display: flex;
+  width: 100%;
+  font-size: 12px;
+  text-align: center;
+  line-height: 2.2rem;
+}
+
+nav a.router-link-exact-active {
+  color: var(--color-text);
+}
+
+nav a.router-link-exact-active:hover {
+  background-color: transparent;
+}
+
+nav a {
+  display: inline-block;
+  padding: 0 1rem;
+  border-left: 1px solid var(--color-border);
+}
+
+nav a:first-of-type {
+  border: 0;
+}
+
+.combo
+{
+  display: table-cell;
+  text-align: center; /* Horizontally center */
+  vertical-align: middle; /* Vertically center */
+}
+
+.loginpage {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  padding: 0 rem;
+  margin: 0;
+  height: 100vh;
+  width: 100vw;
+}
+
+.button {
+  background-color: #c2fbd7;
+  border-radius: 100px;
+  box-shadow: rgba(44, 187, 99, .2) 0 -25px 18px -14px inset,rgba(44, 187, 99, .15) 0 1px 2px,rgba(44, 187, 99, .15) 0 2px 4px,rgba(44, 187, 99, .15) 0 4px 8px,rgba(44, 187, 99, .15) 0 8px 16px,rgba(44, 187, 99, .15) 0 16px 32px;
+  color: green;
+  cursor: pointer;
+  display: inline-block;
+  font-family: CerebriSans-Regular,-apple-system,system-ui,Roboto,sans-serif;
+  padding: 7px 20px;
+  text-align: center;
+  text-decoration: none;
+  transition: all 250ms;
+  border: 0;
+  font-size: 16px;
+  user-select: none;
+  -webkit-user-select: none;
+  touch-action: manipulation;
+  width: 100px;
+  height: 40px;
+  margin-top: 1rem;
+}
+
+.button:hover {
+  box-shadow: rgba(44,187,99,.35) 0 -25px 18px -14px inset,rgba(44,187,99,.25) 0 1px 2px,rgba(44,187,99,.25) 0 2px 4px,rgba(44,187,99,.25) 0 4px 8px,rgba(44,187,99,.25) 0 8px 16px,rgba(44,187,99,.25) 0 16px 32px;
+  transform: scale(1.05) rotate(-1deg);
+}
+
+
+@media (min-width: 1024px) {
+  header {
+    place-items: center;
+    padding-right: 0;
+  }
+
+  nav {
+    flex-direction: column;
+    text-align: left;
+    margin-left: 0rem;
+    font-size: 1.5rem;
+  }
+
+}
 </style>
-
-
