+++ frontend/src/assets/main.css
@@ -1,5 +1,37 @@
 @import './base.css';
 
 #app {
+  margin: 0 auto;
+  padding: 2rem;
 
+  font-weight: normal;
+}
+
+a,
+.green {
+  text-decoration: none;
+  color: hsla(160, 100%, 37%, 1);
+  transition: 0.4s;
+}
+
+@media (hover: hover) {
+  a:hover {
+    background-color: hsla(160, 100%, 37%, 0.2);
+  }
+}
+
+@media (min-width: 1024px) {
+  body {
+    display: flex;
+    place-items: center;
+  }
+
+  #app {
+    display: grid;
+    grid-template-columns: 1fr 6fr;
+    padding: 0 1rem;
+    margin: 0;
+    height: 100vh;
+    width: 100vw;
+  }
 }
