+++ frontend/src/components/Login.vue
@@ -1,0 +1,13 @@
+<script setup lang="ts">
+
+</script>
+
+<template>
+  <div>
+    <h1>Welcome to Raquetas</h1>
+  </div>
+</template>
+
+<style>
+
+</style>
