+++ frontend/src/views/LeaderboardView.vue
